seperated the questions on each component page. Added a nav bar to help see each page as it is coded. Wrapped the app in the Broswer Router
