+++ src/views/Questions/Questions1.tsx
@@ -1,73 +1,122 @@
-// import "./Questions.css";
-// import { useState, ChangeEvent, FormEvent } from "react";
+import "./Questions.css";
+import { useState, useEffect, ChangeEvent, FormEvent } from "react";
 
-// const Questions = () => {
-//   const [formData, setFormData] = useState({
-//     question1: "",
-//     question2: "",
-//     question3: "",
-//     question4: "",
-//     question5: "",
-//     question6: "",
-//     question7: "",
-//     question8: "",
-//     question9: "",
-//     question10: "",
-//   });
+const Questions1 = () => {
+  const [formData, setFormData] = useState({
+    question1: "",
+    question2: "",
+    question3: "",
+    question4: "",
+  });
 
-//   const handleFormSubmit = (event: FormEvent) => {
-//     event.preventDefault();
-//     console.log(formData);
-//   };
+   // Creating different state variables for the error instead of and object
+   const [question1Error, setQuestion1Error] = useState(false);
+   const [question2Error, setQuestion2Error] = useState(false);
+   const [question3Error, setQuestion3Error] = useState(false);
+   const [question4Error, setQuestion4Error] = useState(false);
 
-//   const handleChange = (event: ChangeEvent) => {
-//     const { name, value } = event.target;
-//     setFormData((prevFormData) => ({
-//       ...prevFormData,
-//       [name]: value,
-//     }));
-//   };
-//   return (
-//     <>
-//       <div className="form-container">
-//         <form onSubmit={handleFormSubmit}>
-//           <label className="question-labels">Questions 1:</label>
-//           <input
-//             type="text"
-//             name="question1"
-//             value={formData.question1}
-//             onChange={handleChange}
-//           />
+   useEffect(() => {
+    // Load cached form data if available
+    const cachedFormData = JSON.parse(localStorage.getItem("formData") || "{}");
+    setFormData(cachedFormData);
+  }, []);
 
-//           <label className="question-labels">Question 2:</label>
-//           <input
-//             type="text"
-//             name="question2"
-//             value={formData.question2}
-//             onChange={handleChange}
-//           />
+  const handleFormSubmit = (event: FormEvent) => {
+    event.preventDefault();
+    console.log(formData);
 
-//           <label className="question-labels">Question 3:</label>
+    // Save form data to localStorage
+    localStorage.setItem("formData", JSON.stringify(formData));
 
-//           <label className="question-labels">Question 4:</label>
+    // validate form input
+    let isFormValid = true;
 
-//           <label className="question-labels">Question 5:</label>
+    if (!formData.question1) {
+      isFormValid = false;
+      setQuestion1Error(true);
+    }
 
-//           <label className="question-labels">Question 6:</label>
+    if (!formData.question2) {
+      isFormValid = false;
+      setQuestion2Error(true);
+    }
 
-//           <label className="question-labels">Question 7:</label>
+    if (!formData.question3) {
+      isFormValid = false;
+      setQuestion3Error(true);
+    }
 
-//           <label className="question-labels">Question 8:</label>
+    if (!formData.question4) {
+      isFormValid = false;
+      setQuestion4Error(true);
+    }
 
-//           <label className="question-labels">Question 9:</label>
 
-//           <label className="question-labels">Question 10:</label>
+    if (isFormValid) {
+      //Form submission logic here
+      alert("Form submitted successfully!");
+    } else {
+      alert("Form submission failed. Please fill in all required fields.");
+    }
+    setFormData({
+      question1: "",
+      question2: "",
+      question3: "",
+      question4: "",
+    });
+    };
 
-//           <button type="submit">Submit</button>
-//         </form>
-//       </div>
-//     </>
-//   );
-// };
+  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = event.target;
+    setFormData(({ ...formData, [name]: value }));
+  };
 
-// export default Questions;
+  return (
+    <>
+    <div className="form-container">
+      <h4>Please tell us about your home</h4>
+        <form onSubmit={handleFormSubmit}>
+          <label className="question-labels">Questions 1:</label>
+          <input
+            type="text"
+            name="question1"
+            value={formData.question1}
+            onChange={handleChange}
+          />
+         {question1Error && <div className="error-message">Answer is required.</div>}
+
+          <label className="question-labels">Question 2:</label>
+          <input
+            type="text"
+            name="question2"
+            value={formData.question2}
+            onChange={handleChange}
+          />
+{question2Error && <div className="error-message">Answer is required.</div>}
+
+          <label className="question-labels">Question 3:</label>
+          <input
+            type="text"
+            name="question3"
+            value={formData.question3}
+            onChange={handleChange}
+          />
+          { question3Error && <div className="error-message">Answer is required.</div>}
+
+          <label className="question-labels">Question 4:</label>
+          <input
+            type="text"
+            name="question4"
+            value={formData.question4}
+            onChange={handleChange}
+          />
+          {question4Error && <div className="error-message">Answer is required.</div>}
+          
+          <button className="form-btn">Next</button>
+        </form>
+      </div>
+      </>
+  )
+}
+export default Questions1
+
